backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.123.5",
    "sqlalchemy>=2.0.44",
    "strawberry-graphql[fastapi]>=0.287.2",
]
